rowblk: reorganize tests
Split the `rowblk` tests into multiple files.
